Moving URIs to the URID crate and the LV2 Mapping features to the core crate
