+++ core/src/plugin/info.rs
@@ -1,4 +1,4 @@
-use crate::Uri;
+use lv2_urid::Uri;
 use std::os::raw::c_char;
 use std::path::Path;
 use std::str::Utf8Error;
