+++ core/src/prelude.rs
@@ -1,7 +1,9 @@
 //! Prelude for wildcard use, containing many important types.
-pub use crate::feature::{FeatureCache, FeatureCollection, MissingFeatureError, ThreadingClass};
+pub use crate::feature::{
+    FeatureCache, FeatureCollection, HardRTCapable, InPlaceBroken, IsLive, Map,
+    MissingFeatureError, ThreadingClass, Unmap,
+};
 pub use crate::match_extensions;
 pub use crate::plugin::{lv2_descriptors, Plugin, PluginInfo, PortCollection};
 pub use crate::port::*;
 pub use crate::sys::LV2_Descriptor;
-pub use crate::{Uri, UriBound};
