+++ core/tests/amp.rs
@@ -1,6 +1,7 @@
 use lv2_core::feature::{FeatureCache, FeatureCollection, MissingFeatureError};
 use lv2_core::feature::{HardRTCapable, IsLive};
 use lv2_core::prelude::*;
+use lv2_urid::prelude::*;
 use std::ops::Drop;
 use std::os::raw::c_char;
 
@@ -100,8 +101,8 @@ fn test_discovery() {
 
 #[test]
 fn test_plugin() {
-    use lv2_core::UriBound;
     use lv2_sys::*;
+    use lv2_urid::UriBound;
 
     // Creating the ports.
     let mut gain: f32 = 2.0;
