+++ urid/Cargo.toml
@@ -14,10 +14,9 @@ travis-ci = { repository = "RustAudio/rust-lv2", branch = "master" }
 maintenance = { status = "passively-maintained" }
 
 [dependencies]
-lv2-core = "1.0.0"
 lv2-sys = "1.0.0"
 lv2-urid-derive = "1.0.0"
 
 [features]
 default = []
-host = ["lv2-core/host"]
+host = []
